-EmailVerification

+++ src/main/java/ua/smartsub/smartsub/dao/CommentsDao.java
@@ -3,5 +3,5 @@ package ua.smartsub.smartsub.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.smartsub.smartsub.entity.Comments;
 
-public interface CommentsDao extends  JpaRepository<Comments,Integer> {
+public interface CommentsDao extends  JpaRepository<Comments,Long> {
 }
