+++ src/main/java/ua/smartsub/smartsub/dao/RoleDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ua.smartsub.smartsub.entity.Role;
 
-public interface RoleDao  extends JpaRepository<Role,Integer> {
+public interface RoleDao  extends JpaRepository<Role,Long> {
 
     @Query("select r from Role r where r.name=:name")
     Role findByName(String name);
