+++ src/main/java/ua/smartsub/smartsub/dao/SubscribeDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ua.smartsub.smartsub.entity.Subscribe;
 
-public interface SubscribeDao extends JpaRepository<Subscribe,Integer> {
+public interface SubscribeDao extends JpaRepository<Subscribe,Long> {
 
     @Query("select s from Subscribe s where s.id=:id ")
     Subscribe findById(int id);
